problemes express arreglats

+++ Docker/mediaServer/init.sql
@@ -1,7 +1,7 @@
 use mediaServer
 
 CREATE TABLE video (
-    id INT PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     titol VARCHAR(50),
     duracio FLOAT,
     codec VARCHAR(10),
